requirements.txt
langchain
weaviate-client
openai
python-dotenv
sentence-transformers
transformers
beautifulsoup4
tqdm
requests
seleniumbase
crawl4ai